% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{get_peptides}
\alias{get_peptides}
\title{Get all peptides from epFASTA object as a vector}
\usage{
get_peptides(x = NULL)
}
\arguments{
\item{x}{an epFASTA data object}
}
\value{
a vector
}
\description{
\code{get_peptides()} will return all peptides as string vector
}
