% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.epfasta}
\alias{print.epfasta}
\title{Tidy-Quant data object print definition}
\usage{
\method{print}{epfasta}(x, ...)
}
\arguments{
\item{x}{epfasta data object}

\item{...}{unused legacy}
}
\value{
print object summary
}
\description{
Tidy-Quant data object print definition
}
